refactor database operations for improved clarity and flexibility
